Extract api mocks
